requirements.txt
alembic
psycopg2-binary
python-dotenv
python-telegram-bot
requests==2.26.0
SQLAlchemy
